clean.sh - добавил проверку на наличие файла

+++ scripts/clean.sh
@@ -1,9 +1,12 @@
 for f in $*
 do
-	if file $f | grep -q CRLF; then
-		perl -i -pe 's/\s+$/\r\n/g' $f
-	else
-		perl -i -pe 's/\s+$/\n/g' $f
+	if [ -a $f ]
+	then
+		if file $f | grep -q CRLF; then
+			perl -i -pe 's/\s+$/\r\n/g' $f
+		else
+			perl -i -pe 's/\s+$/\n/g' $f
+		fi
+		rm $f.bak
 	fi
-rm $f.bak
 done
